test: 修正测试目录与功能分类
