+++ scripts/check_deleted_dir_refs.py
@@ -24,6 +24,7 @@ DELETED_DIRS: tuple[str, ...] = (
     "tests/algorithms",
     "tests/numerical/dynamics",
     "tests/data/frames/fixtures",
+    "tests/data/atmosphere",
 )
 
 # 匹配 "tests/<dir>" 后跟 /、引号、空白或行尾；不匹配 tests/core_xxx 之类的延续
