+++ tests/algorithm/dynamics/test_dynamics_contract.py
@@ -20,17 +20,6 @@ def sample_state():
     return np.array([0.8, 0.0, 0.0, 0.0, 0.1, 0.0])
 
 
-def _periodic_orbit():
-    from e2m2e.data.types.orbit import Orbit
-
-    orbit = Orbit(
-        states=np.array([[0.8, 0.0, 0.0, 0.0, 0.1, 0.0]]),
-        times=np.array([0.0]),
-    )
-    orbit.period = 1.0
-    return orbit
-
-
 def test_dynamics_initializes_the_default_integrator_contract(dynamics):
     assert isinstance(dynamics, CR3BP_Dynamics)
     assert dynamics.rtol == 1e-12
@@ -83,20 +72,3 @@ def test_collision_detection_requires_an_explicit_backend(dynamics, sample_state
 def test_collision_detection_is_disabled_by_default(dynamics, sample_state):
     result = dynamics.propagate(sample_state, (0.0, 1.0))
     assert "collision" not in result
-
-
-def test_orbit_state_at_initial_epoch_matches_the_input_state(earth_moon_dynamics):
-    orbit = _periodic_orbit()
-
-    state = earth_moon_dynamics.propagate_orbit_state_at_time(orbit, 0.0)
-
-    assert_allclose(state, orbit.states[0], rtol=1e-9, atol=1e-12)
-
-
-def test_orbit_state_at_time_returns_a_finite_state(earth_moon_dynamics):
-    state = earth_moon_dynamics.propagate_orbit_state_at_time(
-        _periodic_orbit(), 0.05, integration_dt=0.005
-    )
-
-    assert state.shape == (6,)
-    assert np.all(np.isfinite(state))
