--- tests/data/atmosphere/__init__.py
@@ -1,1 +1,0 @@
-"""e2m2e.algorithm.forces.atmosphere 模块测试包。"""
